refactor : viewerpage 중복 제거
