+++ Frontend-main/src/components/graph/RelationGraph_Graphpage.jsx
@@ -614,7 +614,7 @@ function StandaloneRelationGraph({
                     mode={inViewer ? 'viewer' : 'standalone'}
                     chapterNum={chapterNum}
                     eventNum={eventNum}
-                    maxChapter={inViewer ? maxChapter : 10}
+                    maxChapter={maxChapter}
                     style={{ pointerEvents: 'auto' }}
                   />
                 )}
